few improvements
+ move Payment Method classes into Service folder
+ fix imports that can be usable from user
+ fix an issue with getMethod
