+++ .editorconfig
@@ -3,7 +3,7 @@ charset = utf-8
 end_of_line = crlf
 indent_size = 4
 indent_style = space
-insert_final_newline = true
+insert_final_newline = false
 max_line_length = 200
 tab_width = 4
 trim_trailing_whitespace = true
