+++ src/PaymentMethods/AfterpayDigiAccept/index.ts
@@ -1,6 +1,6 @@
 import { IPay } from '../Afterpay/Model/Pay';
 import IRequest, { IRefundRequest } from '../../Models/IRequest';
-import PayablePaymentMethod from '../PayablePaymentMethod';
+import PayablePaymentMethod from '../../Services/PayablePaymentMethod';
 import { Pay } from './Model/Pay';
 
 export default class AfterpayDigiAccept extends PayablePaymentMethod {
