+++ src/PaymentMethods/EPS/index.ts
@@ -1,4 +1,4 @@
-import PayablePaymentMethod from '../PayablePaymentMethod';
+import PayablePaymentMethod from '../../Services/PayablePaymentMethod';
 
 export default class EPS extends PayablePaymentMethod {
     protected _paymentName = 'EPS';
