+++ src/PaymentMethods/Mbway/index.ts
@@ -1,4 +1,4 @@
-import PayablePaymentMethod from '../PayablePaymentMethod';
+import PayablePaymentMethod from '../../Services/PayablePaymentMethod';
 
 export default class Mbway extends PayablePaymentMethod {
     protected _paymentName = 'MB WAY';
