+++ src/PaymentMethods/PaymentMethods.ts
@@ -1,0 +1,84 @@
+export { default as ideal } from './Ideal';
+export { default as idealprocessing } from './Ideal';
+export { default as afterpay } from './Afterpay';
+export { default as afterpaydigiaccept } from './AfterpayDigiAccept';
+export { default as applepay } from './ApplePay';
+export { default as bancontactmrcash } from './Bancontact';
+export { default as transfer } from './BankTransfer';
+export { default as belfius } from './Belfius';
+export { default as billink } from './Billink';
+export { default as buckaroovoucher } from './BuckarooVoucher';
+export { default as BuckarooWalletCollecting } from './BuckarooWallet';
+export { default as CreditCard } from './CreditCard';
+
+// Credit Cards
+export { default as creditcard } from './CreditCard';
+export { default as mastercard } from './CreditCard';
+export { default as visa } from './CreditCard';
+export { default as amex } from './CreditCard';
+export { default as vpay } from './CreditCard';
+export { default as maestro } from './CreditCard';
+export { default as visaelectron } from './CreditCard';
+export { default as cartebleuevisa } from './CreditCard';
+export { default as cartebancaire } from './CreditCard';
+export { default as dankort } from './CreditCard';
+export { default as nexi } from './CreditCard';
+export { default as postepay } from './CreditCard';
+
+export { default as creditclick } from './CreditClick';
+export { default as CreditManagement3 } from './CreditManagement';
+export { default as emandate } from './Emandates';
+export { default as eps } from './EPS';
+
+// Gift Cards
+export { default as giftcard } from './GiftCard';
+export { default as westlandbon } from './GiftCard';
+export { default as babygiftcard } from './GiftCard';
+export { default as babyparkgiftcard } from './GiftCard';
+export { default as beautywellness } from './GiftCard';
+export { default as boekenbon } from './GiftCard';
+export { default as boekenvoordeel } from './GiftCard';
+export { default as designshopsgiftcard } from './GiftCard';
+export { default as fashioncheque } from './GiftCard';
+export { default as fashionucadeaukaart } from './GiftCard';
+export { default as fijncadeau } from './GiftCard';
+export { default as koffiecadeau } from './GiftCard';
+export { default as kokenzo } from './GiftCard';
+export { default as kookcadeau } from './GiftCard';
+export { default as nationaleentertainmentcard } from './GiftCard';
+export { default as naturesgift } from './GiftCard';
+export { default as podiumcadeaukaart } from './GiftCard';
+export { default as shoesaccessories } from './GiftCard';
+export { default as webshopgiftcard } from './GiftCard';
+export { default as wijncadeau } from './GiftCard';
+export { default as wonenzo } from './GiftCard';
+export { default as yourgift } from './GiftCard';
+export { default as vvvgiftcard } from './GiftCard';
+export { default as parfumcadeaukaart } from './GiftCard';
+
+export { default as giropay } from './Giropay';
+export { default as idealqr } from './IdealQR';
+export { default as idin } from './Idin';
+export { default as capayable } from './In3Old';
+export { default as KBCPaymentButton } from './KBC';
+export { default as klarna } from './Klarna';
+export { default as klarnakp } from './KlarnaKP';
+export { default as Marketplaces } from './Marketplaces';
+export { default as payconiq } from './Payconiq';
+export { default as PayByBank } from './PayByBank';
+export { default as paypal } from './Paypal';
+export { default as payperemail } from './PayPerEmail';
+export { default as PiM } from './PiM';
+export { default as pointofsale } from './PointOfSale';
+export { default as przelewy24 } from './Przelewy24';
+export { default as sepadirectdebit } from './SEPA';
+export { default as sofortueberweisung } from './Sofort';
+export { default as subscriptions } from './Subscriptions';
+export { default as surepay } from './Surepay';
+export { default as thunes } from './Thunes';
+export { default as tinka } from './Tinka';
+export { default as alipay } from './Alipay';
+export { default as trustly } from './Trustly';
+export { default as wechatpay } from './WeChatPay';
+export { default as In3 } from './In3';
+export { default as noservice } from './NoService';
