+++ src/Utils/MethodTypes.ts
@@ -1,11 +1,11 @@
-import * as AllPaymentMethods from '../PaymentMethods';
+import * as AllPaymentMethods from '../PaymentMethods/PaymentMethods';
 
 export type AvailablePaymentMethods = typeof AllPaymentMethods;
 export type ServiceCode = keyof AvailablePaymentMethods;
 export type PaymentMethodInstance<Code extends ServiceCode> = InstanceType<AvailablePaymentMethods[Code]>;
 
 export function getMethod<Code extends ServiceCode>(code: Code): PaymentMethodInstance<Code> {
-    const methodClass = AllPaymentMethods['ideal'];
+    const methodClass = AllPaymentMethods[code];
     if (!methodClass) {
         throw new Error(`Invalid payment method code: ${code}`);
     }
