+++ src/Utils/index.ts
@@ -1,0 +1,3 @@
+export * from './MethodTypes';
+export * from './Types';
+export { getIPAddress, DataFormatter } from './Functions';
